# docs/specs/patient-service-openapi.yaml
openapi: 3.0.0
info:
  title: Patient Service API
  version: 1.0.0
paths:
...
  /appointments:
    post:
      summary: Book appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentDTO'
      responses:
        '201':
          description: Created
  /appointments/patient/{id}:
    get:
      summary: Get appointments by patient ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
...
components:
  schemas:
...
    AppointmentDTO:
      type: object
      properties:
        patientId: { type: integer }
        doctorId: { type: integer }
        dateTime: { type: string, format: date-time }